feat: new listeners & inSlot condition, cooldown reappear (#49)
* feat: new listeners & inSlot condition

Adds the following observable events:

- itemLeftClickSneak: sneak + left click
- itemRightClickSneak: sneak + right click
- itemSwapIn: switch into item or swap hands
- itemSwapOut: switch out of active item or swap hands
- itemSneak: crouch while item is equipped

This allows different actions to run on right click
and on sneak+right click, for example.

Adds the inSlot condition:

For example:

- ensure:
  inSlot:
    slots: [9, 10, 38] - passive slots + chestplate slot
    item: mineinabyss:test_passive_1 - should usually be the same item but not required to be

Finally, several new slots will receive events.
These are the 4 equipment slots: helmet, chestplate, leggings, and boots, as well
as slots 9 and 10 (which are the 2 top-left-most slots).
The active hand slot will not receive the itemSneak event.

* fix: item/equipment split, offhand swaps

Events have been split, now there are equipment
events for items in the equipment slots and item events for
items in the mainhand.

* feat: cooldown reappears on ability activation